+++ firmware/esp8266/libraries/ESPway/examples/ClockDemo/ClockDisplay.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <Arduino.h>
+#include <time.h>
+
+class ClockDisplay {
+ public:
+  virtual ~ClockDisplay() = default;
+  virtual void begin() = 0;
+  virtual void show(const tm* localTime, bool valid, bool ntpMode, bool synced) = 0;
+};